refactor kotlin timeSpent data type

+++ kotlin-root/gasnake/src/main/kotlin/com/example/snake/ai/ga/EvolveResult.kt
@@ -6,7 +6,7 @@ import com.example.snake.ai.BaseStats
 data class EvolveResult(
     val generation: Int,
     val bestIndividual: IndividualPlainObject,
-    val timeSpent: Double,
+    val timeSpent: Long,
     val overallStats: OverallStats
 )
